requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pytesseract==0.3.13
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
